mcp-server/requirements.txt
# NEXUS MCP Server dependencies
mcp[cli]>=1.9.0
fastmcp>=2.0.0
httpx>=0.27.0
pydantic>=2.0.0
uvicorn>=0.30.0
google-auth>=2.29.0
google-api-python-client>=2.130.0
fastapi>=0.115.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0

# v2 update: browser automation (optional — install chromium separately)
# To enable browser_* tools, also run: playwright install chromium --with-deps
playwright>=1.49.0
pytest>=8.3.0
